RadioButtonContainer class moved to models
