+++ src/main/java/wrappers/RadioButtonInterface.java
@@ -1,5 +1,6 @@
 package wrappers;
 
+import models.RadioButtonContainer;
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
@@ -54,29 +55,5 @@ public class RadioButtonInterface {
                 .orElseThrow(() -> new NoSuchElementException("No radio button found with provided value"))
                 .getInput()
                 .click();
-    }
-
-    public static class RadioButtonContainer {
-        private final UIElement label;
-        private final UIElement input;
-        private final UIElement description;
-
-        public RadioButtonContainer(UIElement label, UIElement input, UIElement description) {
-            this.label = label;
-            this.input = input;
-            this.description = description;
-        }
-
-        public UIElement getLabel() {
-            return label;
-        }
-
-        public UIElement getInput() {
-            return input;
-        }
-
-        public UIElement getDescription() {
-            return description;
-        }
     }
 }
